data_importer/requirements.txt
pandas
sqlalchemy
psycopg2-binary
werkzeug
python-dotenv